AKS/AML Dependent/src/requirement.txt
applicationinsights
azureml-defaults>=1.0.45
azure-storage-blob
azure-identity
pandas
scikit-learn==0.22.2.post1